+++ NEWS.md
@@ -1,1 +1,4 @@
 # purrr 0.1.0.9000
+
+* Implement `set_names()`: a camel case alternative to `setNames()` with
+  stricter equality checking (#119).